requirements.txt
eventlet==0.33.1
python-socketio==5.7.2